% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_new.R
\name{segment_new}
\alias{segment_new}
\title{Segmenting the numeric field}
\usage{
segment_new(access_token, instance_url, object, field)
}
\description{
This function build a categorical variable from a numeric variable
}
