% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\name{updater}
\alias{updater}
\title{Updating Salesforce}
\usage{
updater(access_token, instance_url, myobject, data)
}
\arguments{
\item{access_token}{for salesforce credentials}

\item{instance_url}{is salesforce url}

\item{myobject}{is object name}

\item{data}{is dataset value}
}
\description{
This function is used to update salesforce objects with the current result
}
